+++ server/src/main/java/org/opensearch/extensions/ExtensionsOrchestrator.java
@@ -73,6 +73,9 @@ public class ExtensionsOrchestrator implements ReportingService<PluginsAndModule
     public static final String REQUEST_EXTENSION_CLUSTER_STATE = "internal:discovery/clusterstate";
     public static final String REQUEST_EXTENSION_LOCAL_NODE = "internal:discovery/localnode";
     public static final String REQUEST_EXTENSION_CLUSTER_SETTINGS = "internal:discovery/clustersettings";
+    public static final String REQUEST_EXTENSION_ENVIRONMENT_SETTINGS = "internal:discovery/enviornmentsettings";
+    public static final String REQUEST_EXTENSION_ADD_SETTINGS_UPDATE_CONSUMER = "internal:discovery/addsettingsupdateconsumer";
+    public static final String REQUEST_EXTENSION_UPDATE_SETTINGS = "internal:discovery/updatesettings";
     public static final String REQUEST_EXTENSION_REGISTER_CUSTOM_SETTINGS = "internal:discovery/registercustomsettings";
     public static final String REQUEST_EXTENSION_REGISTER_REST_ACTIONS = "internal:discovery/registerrestactions";
     public static final String REQUEST_EXTENSION_REGISTER_TRANSPORT_ACTIONS = "internal:discovery/registertransportactions";
@@ -121,6 +124,8 @@ public class ExtensionsOrchestrator implements ReportingService<PluginsAndModule
     ExtensionNamedWriteableRegistry namedWriteableRegistry;
     ExtensionActionListener listener;
     ExtensionActionListenerHandler listenerHandler;
+    EnvironmentSettingsRequestHandler environmentSettingsRequestHandler;
+    AddSettingsUpdateConsumerRequestHandler addSettingsUpdateConsumerRequestHandler;
 
     /**
      * Instantiate a new ExtensionsOrchestrator object to handle requests and responses from extensions. This is called during Node bootstrap.
@@ -147,25 +152,33 @@ public class ExtensionsOrchestrator implements ReportingService<PluginsAndModule
     }
 
     /**
-     * Initializes the {@link RestActionsRequestHandler}, {@link TransportService} and {@link ClusterService}. This is called during Node bootstrap.
+     * Initializes the {@link RestActionsRequestHandler}, {@link TransportService}, {@link ClusterService} and environment settings. This is called during Node bootstrap.
      * Lists/maps of extensions have already been initialized but not yet populated.
      *
      * @param restController  The RestController on which to register Rest Actions.
      * @param settingsModule Settings Module
      * @param transportService  The Node's transport service.
      * @param clusterService  The Node's cluster service.
+     * @param initialEnvironmentSettings The finalized view of settings for the Environment
      */
     public void initializeServicesAndRestHandler(
         RestController restController,
         SettingsModule settingsModule,
         TransportService transportService,
-        ClusterService clusterService
+        ClusterService clusterService,
+        Settings initialEnvironmentSettings
     ) {
         this.restActionsRequestHandler = new RestActionsRequestHandler(restController, extensionIdMap, transportService);
         this.listenerHandler = new ExtensionActionListenerHandler(listener);
         this.customSettingsRequestHandler = new CustomSettingsRequestHandler(settingsModule);
         this.transportService = transportService;
         this.clusterService = clusterService;
+        this.environmentSettingsRequestHandler = new EnvironmentSettingsRequestHandler(initialEnvironmentSettings);
+        this.addSettingsUpdateConsumerRequestHandler = new AddSettingsUpdateConsumerRequestHandler(
+            clusterService,
+            transportService,
+            REQUEST_EXTENSION_UPDATE_SETTINGS
+        );
         registerRequestHandler();
     }
 
@@ -217,6 +230,24 @@ public class ExtensionsOrchestrator implements ReportingService<PluginsAndModule
             false,
             ExtensionActionListenerOnFailureRequest::new,
             ((request, channel, task) -> channel.sendResponse(listenerHandler.handleExtensionActionListenerOnFailureRequest(request)))
+        );
+        transportService.registerRequestHandler(
+            REQUEST_EXTENSION_ENVIRONMENT_SETTINGS,
+            ThreadPool.Names.GENERIC,
+            false,
+            false,
+            EnvironmentSettingsRequest::new,
+            ((request, channel, task) -> channel.sendResponse(environmentSettingsRequestHandler.handleEnvironmentSettingsRequest(request)))
+        );
+        transportService.registerRequestHandler(
+            REQUEST_EXTENSION_ADD_SETTINGS_UPDATE_CONSUMER,
+            ThreadPool.Names.GENERIC,
+            false,
+            false,
+            AddSettingsUpdateConsumerRequest::new,
+            ((request, channel, task) -> channel.sendResponse(
+                addSettingsUpdateConsumerRequestHandler.handleAddSettingsUpdateConsumerRequest(request)
+            ))
         );
         transportService.registerRequestHandler(
             REQUEST_EXTENSION_REGISTER_TRANSPORT_ACTIONS,
@@ -349,6 +380,22 @@ public class ExtensionsOrchestrator implements ReportingService<PluginsAndModule
     }
 
     /**
+     * Handles a {@link RegisterTransportActionsRequest}.
+     *
+     * @param transportActionsRequest  The request to handle.
+     * @return  A {@link ExtensionBooleanResponse} indicating success.
+     * @throws Exception if the request is not handled properly.
+     */
+    TransportResponse handleRegisterTransportActionsRequest(RegisterTransportActionsRequest transportActionsRequest) throws Exception {
+        /*
+         * TODO: https://github.com/opensearch-project/opensearch-sdk-java/issues/107
+         * Register these new Transport Actions with ActionModule
+         * and add support for NodeClient to recognise these actions when making transport calls.
+         */
+        return new ExtensionBooleanResponse(true);
+    }
+
+    /**
      * Handles an {@link ExtensionRequest}.
      *
      * @param extensionRequest  The request to handle, of a type defined in the {@link RequestType} enum.
@@ -366,22 +413,6 @@ public class ExtensionsOrchestrator implements ReportingService<PluginsAndModule
             default:
                 throw new IllegalArgumentException("Handler not present for the provided request");
         }
-    }
-
-    /**
-     * Handles a {@link RegisterTransportActionsRequest}.
-     *
-     * @param transportActionsRequest  The request to handle.
-     * @return  A {@link ExtensionBooleanResponse} indicating success.
-     * @throws Exception if the request is not handled properly.
-     */
-    TransportResponse handleRegisterTransportActionsRequest(RegisterTransportActionsRequest transportActionsRequest) throws Exception {
-        /*
-         * TODO: https://github.com/opensearch-project/opensearch-sdk-java/issues/107
-         * Register these new Transport Actions with ActionModule
-         * and add support for NodeClient to recognise these actions when making transport calls.
-         */
-        return new ExtensionBooleanResponse(true);
     }
 
     public void onIndexModule(IndexModule indexModule) throws UnknownHostException {