+++ server/src/main/java/org/opensearch/node/Node.java
@@ -778,7 +778,13 @@ public class Node implements Closeable {
              * TODO: Understand the dependencies from plugins to initialize TransportService.
              *  This seems like a chicken and egg problem.
              */
-            this.extensionsOrchestrator.initializeServicesAndRestHandler(restController, settingsModule, transportService, clusterService);
+            this.extensionsOrchestrator.initializeServicesAndRestHandler(
+                restController,
+                settingsModule,
+                transportService,
+                clusterService,
+                environment.settings()
+            );
             final GatewayMetaState gatewayMetaState = new GatewayMetaState();
             final ResponseCollectorService responseCollectorService = new ResponseCollectorService(clusterService);
             final SearchTransportService searchTransportService = new SearchTransportService(