+++ CHANGELOG.md
@@ -101,6 +101,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
  - Handle named wildcards (REST path parameters) ([#4415](https://github.com/opensearch-project/OpenSearch/pull/4415))
  - Adding ActionListener onFailure to ExtensionsOrchestrator ([#4191](https://github.com/opensearch-project/OpenSearch/pull/4191))
  - Created new Request class for ActionListener onFailure ([#4553](https://github.com/opensearch-project/OpenSearch/pull/4553))
+ - Adding create component extension point support for AD ([#4517](https://github.com/opensearch-project/OpenSearch/pull/4517))
  - Add getSettings support for AD([#4519](https://github.com/opensearch-project/OpenSearch/pull/4519))
 
 ## [2.x]